Fix #3343
Right after leaving Armature editmode, and entering posemode, the active
bone was not selected (when only the head or tip was selected). This caused
a confusing display of an active but not selected Bone... and not proper
behaviour of UI (buttons) and inserting keys.

Fix #3348

Mirror Modifier, button "Merge Limit" didn't react on left/right click.