Bugfix #3426
Object duplicators didn't draw when not selected. Was one 'else' to many in
the code!

Also noticed that duplicators couldn't be transformed when they have Ipos,
was an improper call to do_ipo in duplicator code... also made sure the
transformation info in the object is stored/restored before duplicating.